+++ bootstrap_cfn/config.py
@@ -321,8 +321,10 @@ class ConfigParser(object):
             },
         )
         # Set required policy actions
+        # elasticloadbalancing:Describe* "*" needed for aws-formula and cron-formula
         policy_actions = [{"Action": ["autoscaling:Describe*"], "Resource": "*", "Effect": "Allow"},
-                          {"Action": ["cloudformation:Describe*"], "Resource": "*", "Effect": "Allow"}]
+                          {"Action": ["cloudformation:Describe*"], "Resource": "*", "Effect": "Allow"},
+                          {"Action": ["elasticloadbalancing:Describe*"], "Resource": "*", "Effect": "Allow"}]
 
         # Only define policy actions if the components are enabled in the config
         if 'ec2' in self.data:
@@ -331,7 +333,6 @@ class ConfigParser(object):
         if 'rds' in self.data:
             policy_actions.append({"Action": ["rds:Describe*"], "Resource": "*", "Effect": "Allow"})
         if 'elasticache' in self.data:
-            policy_actions.append({"Action": ["elasticloadbalancing:Describe*"], "Resource": "*", "Effect": "Allow"})
             policy_actions.append({"Action": ["elasticache:Describe*"], "Resource": "*", "Effect": "Allow"})
         if 's3' in self.data:
             policy_actions.append({"Action": ["s3:List*"], "Resource": "*", "Effect": "Allow"})